Changed C++ data type of copy_nb from uint32_t to uint8_t

+++ catalogue/TapeFileWritten.hpp
@@ -102,7 +102,7 @@ struct TapeFileWritten: public TapeItemWritten {
   /**
    * The copy number of the tape file.
    */
-  uint32_t copyNb;
+  uint8_t copyNb;
 
 }; // struct TapeFileWritten
