+++ common/dataStructures/ArchiveFile.cpp
@@ -60,7 +60,7 @@ bool ArchiveFile::operator!=(const ArchiveFile &rhs) const {
 //------------------------------------------------------------------------------
 // ArchiveFile::TapeFilesList::at()
 //------------------------------------------------------------------------------
-TapeFile& ArchiveFile::TapeFilesList::at(uint32_t copyNb) {
+TapeFile& ArchiveFile::TapeFilesList::at(uint8_t copyNb) {
   auto ret = std::find_if(begin(), end(), [=](TapeFile& tf){ return tf.isActiveCopyNb(copyNb);});
   if (ret == end()) throw cta::exception::Exception("In ArchiveFile::TapeFilesList::at(): not found.");
   return *ret;
@@ -69,7 +69,7 @@ TapeFile& ArchiveFile::TapeFilesList::at(uint32_t copyNb) {
 //------------------------------------------------------------------------------
 // ArchiveFile::TapeFilesList::at() const
 //------------------------------------------------------------------------------
-const TapeFile& ArchiveFile::TapeFilesList::at(uint32_t copyNb) const {
+const TapeFile& ArchiveFile::TapeFilesList::at(uint8_t copyNb) const {
   auto ret = std::find_if(cbegin(), cend(), [=](const TapeFile& tf){ return tf.isActiveCopyNb(copyNb);});
   if (ret == end()) throw cta::exception::Exception("In ArchiveFile::TapeFilesList::at(): not found.");
   return *ret;
@@ -78,14 +78,14 @@ const TapeFile& ArchiveFile::TapeFilesList::at(uint32_t copyNb) const {
 //------------------------------------------------------------------------------
 // ArchiveFile::TapeFilesList::find() 
 //------------------------------------------------------------------------------
-ArchiveFile::TapeFilesList::iterator ArchiveFile::TapeFilesList::find(uint32_t copyNb) {
+ArchiveFile::TapeFilesList::iterator ArchiveFile::TapeFilesList::find(uint8_t copyNb) {
   return std::find_if(begin(), end(), [=](TapeFile& tf){ return tf.isActiveCopyNb(copyNb);});
 }
 
 //------------------------------------------------------------------------------
 // ArchiveFile::TapeFilesList::find() const
 //------------------------------------------------------------------------------
-ArchiveFile::TapeFilesList::const_iterator ArchiveFile::TapeFilesList::find(uint32_t copyNb) const {
+ArchiveFile::TapeFilesList::const_iterator ArchiveFile::TapeFilesList::find(uint8_t copyNb) const {
   return std::find_if(cbegin(), cend(), [=](const TapeFile& tf){ return tf.isActiveCopyNb(copyNb);});
 }
